Implement a directional field of view
Field of view is now limited to 140 degrees and is stored as a list containing the coordinates of visible tiles, allowing for easier iteration. The player's facing automatically changes to the direction they move in. Pressing shift with a directional key changes facing without moving, while pressing control will move without changing facing. All entities currently keep track of facing, but this will change with the addition of the entity-component system in the next part of the tutorial.

Some changes were made to input handling to better support modifiers. Fullscreen toggle was also remapped to F11.