Implement a scrolling map and optimize rendering
The map is now always centered on the player, meaning that the map size is no longer limited by the screen size. To take advantage of this, I have roughhly doubled the map size.

Previously, the render_all function would iterate through every tile on the screen and check if it was in the FOV or the memory. This became ridiculously expensive computation-wise as the memory list began to grow. Now the memory and FOV are only iterated over once each, greatly improving performance.

Lastly, a few comments have been added to describe some parts of the code, particularly constants.