Add multiple dungeon levels (tutorial part 11)
The experience system I've implemented is much simpler than the one in the tutorial, granting a flat 10 max health per floor. And yes, I'm skipping over part 10 for now since I'm in a bit of a rush. I'd like to finish with the core mechanics before adding additional features like saving and loading.