Implement in-game switching of color and input schemes
As of yet there is no UI indication of the change. Soon I'll add text labels to each of the schemes that will be displayed as messages when switched to.

I've also disabled the angle-limited FOV for now. It could be cool in some other game for stealth or the immersion factor, but for a straightforward roguelike it adds nothing to gameplay.