Dynamically define entity component fields
Components are now stored in a dict, cutting down on arguments and redundant code. This makes adding new component types even easier, requiring only an addition to the Components enum. The setattr() function has been used to define fields with the name of the component, meaning much less code has to be rewritten (only constructors, in this case).